Move all use of fizz::client::ClientStateMachine into FizzClientHandshake
