+++ quic/client/handshake/ClientHandshake.cpp
@@ -211,13 +211,17 @@ void ClientHandshake::computeCiphers(CipherKind kind, folly::ByteRange secret) {
   }
 }
 
+void ClientHandshake::raiseError(folly::exception_wrapper error) {
+  error_ = std::move(error);
+}
+
+void ClientHandshake::waitForData() {
+  waitForData_ = true;
+}
+
 EncryptionLevel ClientHandshake::getReadRecordLayerEncryptionLevel() {
   return getEncryptionLevelFromFizz(
       state_.readRecordLayer()->getEncryptionLevel());
-}
-
-void ClientHandshake::processSocketData(folly::IOBufQueue& queue) {
-  processActions(machine_.processSocketData(state_, queue));
 }
 
 std::unique_ptr<Aead> ClientHandshake::buildAead(
@@ -278,135 +282,6 @@ void ClientHandshake::computeOneRttCipher(bool earlyDataAccepted) {
   // ClientCleartext. We assume that by the time we get the data for the QUIC
   // stream, the server would have also acked all the client initial packets.
   phase_ = Phase::OneRttKeysDerived;
-}
-
-class ClientHandshake::ActionMoveVisitor : public boost::static_visitor<> {
- public:
-  explicit ActionMoveVisitor(ClientHandshake& client) : client_(client) {}
-
-  void operator()(fizz::DeliverAppData&) {
-    client_.error_ = folly::make_exception_wrapper<QuicTransportException>(
-        "Invalid app data on crypto stream",
-        TransportErrorCode::PROTOCOL_VIOLATION);
-  }
-
-  void operator()(fizz::WriteToSocket& write) {
-    for (auto& content : write.contents) {
-      auto encryptionLevel =
-          getEncryptionLevelFromFizz(content.encryptionLevel);
-      client_.writeDataToStream(encryptionLevel, std::move(content.data));
-    }
-  }
-
-  void operator()(fizz::client::ReportEarlyHandshakeSuccess&) {
-    client_.computeZeroRttCipher();
-  }
-
-  void operator()(fizz::client::ReportHandshakeSuccess& handshakeSuccess) {
-    client_.computeOneRttCipher(handshakeSuccess.earlyDataAccepted);
-  }
-
-  void operator()(fizz::client::ReportEarlyWriteFailed&) {
-    LOG(DFATAL) << "QUIC TLS app data write";
-  }
-
-  void operator()(fizz::ReportError& err) {
-    auto errMsg = err.error.what();
-    if (errMsg.empty()) {
-      errMsg = "Error during handshake";
-    }
-
-    auto fe = err.error.get_exception<fizz::FizzException>();
-
-    if (fe && fe->getAlert()) {
-      auto alertNum =
-          static_cast<std::underlying_type<TransportErrorCode>::type>(
-              fe->getAlert().value());
-      alertNum += static_cast<std::underlying_type<TransportErrorCode>::type>(
-          TransportErrorCode::CRYPTO_ERROR);
-      client_.error_ = folly::make_exception_wrapper<QuicTransportException>(
-          errMsg.toStdString(), static_cast<TransportErrorCode>(alertNum));
-    } else {
-      client_.error_ = folly::make_exception_wrapper<QuicTransportException>(
-          errMsg.toStdString(),
-          static_cast<TransportErrorCode>(
-              fizz::AlertDescription::internal_error));
-    }
-  }
-
-  void operator()(fizz::WaitForData&) {
-    client_.waitForData_ = true;
-  }
-
-  void operator()(fizz::client::MutateState& mutator) {
-    mutator(client_.state_);
-  }
-
-  void operator()(fizz::client::NewCachedPsk& newCachedPsk) {
-    if (client_.callback_) {
-      client_.callback_->onNewCachedPsk(newCachedPsk);
-    }
-  }
-
-  void operator()(fizz::EndOfData&) {
-    client_.error_ = folly::make_exception_wrapper<QuicTransportException>(
-        "unexpected close notify", TransportErrorCode::INTERNAL_ERROR);
-  }
-
-  void operator()(fizz::SecretAvailable& secretAvailable) {
-    folly::variant_match(
-        secretAvailable.secret.type,
-        [&](fizz::EarlySecrets earlySecrets) {
-          switch (earlySecrets) {
-            case fizz::EarlySecrets::ClientEarlyTraffic:
-              client_.computeCiphers(
-                  CipherKind::ZeroRttWrite,
-                  folly::range(secretAvailable.secret.secret));
-              break;
-            default:
-              break;
-          }
-        },
-        [&](fizz::HandshakeSecrets handshakeSecrets) {
-          switch (handshakeSecrets) {
-            case fizz::HandshakeSecrets::ClientHandshakeTraffic:
-              client_.computeCiphers(
-                  CipherKind::HandshakeWrite,
-                  folly::range(secretAvailable.secret.secret));
-              break;
-            case fizz::HandshakeSecrets::ServerHandshakeTraffic:
-              client_.computeCiphers(
-                  CipherKind::HandshakeRead,
-                  folly::range(secretAvailable.secret.secret));
-              break;
-          }
-        },
-        [&](fizz::AppTrafficSecrets appSecrets) {
-          switch (appSecrets) {
-            case fizz::AppTrafficSecrets::ClientAppTraffic:
-              client_.computeCiphers(
-                  CipherKind::OneRttWrite,
-                  folly::range(secretAvailable.secret.secret));
-              break;
-            case fizz::AppTrafficSecrets::ServerAppTraffic:
-              client_.computeCiphers(
-                  CipherKind::OneRttRead,
-                  folly::range(secretAvailable.secret.secret));
-              break;
-          }
-        },
-        [&](auto) {});
-  }
-
- private:
-  ClientHandshake& client_;
-};
-
-void ClientHandshake::processActions(fizz::client::Actions actions) {
-  ActionMoveVisitor visitor(*this);
-  for (auto& action : actions) {
-    boost::apply_visitor(visitor, action);
-  }
 }
 
 } // namespace quic
