+++ quic/client/handshake/FizzClientHandshake.cpp
@@ -9,6 +9,7 @@
 #include <quic/client/handshake/FizzClientHandshake.h>
 
 #include <quic/client/handshake/FizzClientQuicHandshakeContext.h>
+#include <quic/handshake/FizzBridge.h>
 #include <quic/handshake/FizzCryptoFactory.h>
 
 namespace quic {
@@ -44,6 +45,139 @@ void FizzClientHandshake::connect(
       std::move(hostname),
       std::move(cachedPsk),
       transportParams));
+}
+
+void FizzClientHandshake::processSocketData(folly::IOBufQueue& queue) {
+  processActions(machine_.processSocketData(state_, queue));
+}
+
+class FizzClientHandshake::ActionMoveVisitor : public boost::static_visitor<> {
+ public:
+  explicit ActionMoveVisitor(FizzClientHandshake& client) : client_(client) {}
+
+  void operator()(fizz::DeliverAppData&) {
+    client_.raiseError(folly::make_exception_wrapper<QuicTransportException>(
+        "Invalid app data on crypto stream",
+        TransportErrorCode::PROTOCOL_VIOLATION));
+  }
+
+  void operator()(fizz::WriteToSocket& write) {
+    for (auto& content : write.contents) {
+      auto encryptionLevel =
+          getEncryptionLevelFromFizz(content.encryptionLevel);
+      client_.writeDataToStream(encryptionLevel, std::move(content.data));
+    }
+  }
+
+  void operator()(fizz::client::ReportEarlyHandshakeSuccess&) {
+    client_.computeZeroRttCipher();
+  }
+
+  void operator()(fizz::client::ReportHandshakeSuccess& handshakeSuccess) {
+    client_.computeOneRttCipher(handshakeSuccess.earlyDataAccepted);
+  }
+
+  void operator()(fizz::client::ReportEarlyWriteFailed&) {
+    LOG(DFATAL) << "QUIC TLS app data write";
+  }
+
+  void operator()(fizz::ReportError& err) {
+    auto errMsg = err.error.what();
+    if (errMsg.empty()) {
+      errMsg = "Error during handshake";
+    }
+
+    auto fe = err.error.get_exception<fizz::FizzException>();
+
+    if (fe && fe->getAlert()) {
+      auto alertNum =
+          static_cast<std::underlying_type<TransportErrorCode>::type>(
+              fe->getAlert().value());
+      alertNum += static_cast<std::underlying_type<TransportErrorCode>::type>(
+          TransportErrorCode::CRYPTO_ERROR);
+      client_.raiseError(folly::make_exception_wrapper<QuicTransportException>(
+          errMsg.toStdString(), static_cast<TransportErrorCode>(alertNum)));
+    } else {
+      client_.raiseError(folly::make_exception_wrapper<QuicTransportException>(
+          errMsg.toStdString(),
+          static_cast<TransportErrorCode>(
+              fizz::AlertDescription::internal_error)));
+    }
+  }
+
+  void operator()(fizz::WaitForData&) {
+    client_.waitForData();
+  }
+
+  void operator()(fizz::client::MutateState& mutator) {
+    mutator(client_.state_);
+  }
+
+  void operator()(fizz::client::NewCachedPsk& newCachedPsk) {
+    if (client_.callback_) {
+      client_.callback_->onNewCachedPsk(newCachedPsk);
+    }
+  }
+
+  void operator()(fizz::EndOfData&) {
+    client_.raiseError(folly::make_exception_wrapper<QuicTransportException>(
+        "unexpected close notify", TransportErrorCode::INTERNAL_ERROR));
+  }
+
+  void operator()(fizz::SecretAvailable& secretAvailable) {
+    folly::variant_match(
+        secretAvailable.secret.type,
+        [&](fizz::EarlySecrets earlySecrets) {
+          switch (earlySecrets) {
+            case fizz::EarlySecrets::ClientEarlyTraffic:
+              client_.computeCiphers(
+                  CipherKind::ZeroRttWrite,
+                  folly::range(secretAvailable.secret.secret));
+              break;
+            default:
+              break;
+          }
+        },
+        [&](fizz::HandshakeSecrets handshakeSecrets) {
+          switch (handshakeSecrets) {
+            case fizz::HandshakeSecrets::ClientHandshakeTraffic:
+              client_.computeCiphers(
+                  CipherKind::HandshakeWrite,
+                  folly::range(secretAvailable.secret.secret));
+              break;
+            case fizz::HandshakeSecrets::ServerHandshakeTraffic:
+              client_.computeCiphers(
+                  CipherKind::HandshakeRead,
+                  folly::range(secretAvailable.secret.secret));
+              break;
+          }
+        },
+        [&](fizz::AppTrafficSecrets appSecrets) {
+          switch (appSecrets) {
+            case fizz::AppTrafficSecrets::ClientAppTraffic:
+              client_.computeCiphers(
+                  CipherKind::OneRttWrite,
+                  folly::range(secretAvailable.secret.secret));
+              break;
+            case fizz::AppTrafficSecrets::ServerAppTraffic:
+              client_.computeCiphers(
+                  CipherKind::OneRttRead,
+                  folly::range(secretAvailable.secret.secret));
+              break;
+          }
+        },
+        [&](auto) {});
+  }
+
+ private:
+  FizzClientHandshake& client_;
+};
+
+void FizzClientHandshake::processActions(fizz::client::Actions actions) {
+  ActionMoveVisitor visitor(*this);
+  for (auto& action : actions) {
+    boost::apply_visitor(visitor, action);
+  }
 }
 
 } // namespace quic
